Move media/starboard headers into media/base to prevent circular dependency
