+++ media/starboard/BUILD.gn
@@ -79,10 +79,7 @@ source_set("starboard") {
   if (use_starboard_media) {
     sources += [
       "bidirectional_fit_reuse_allocator.cc",
-      "bidirectional_fit_reuse_allocator.h",
       "decoder_buffer_allocator.cc",
-      "decoder_buffer_allocator.h",
-      "decoder_buffer_memory_info.h",
       "sbplayer_bridge.cc",
       "sbplayer_bridge.h",
       "sbplayer_interface.cc",
@@ -93,7 +90,6 @@ source_set("starboard") {
       "starboard_cdm.h",
       "starboard_cdm_factory.cc",
       "starboard_cdm_factory.h",
-      "starboard_memory_allocator.h",
       "starboard_renderer.cc",
       "starboard_renderer.h",
       "starboard_renderer_factory.cc",
@@ -116,9 +112,6 @@ source_set("starboard") {
   if (!media_use_ffmpeg) {
     deps += [ "//third_party/abseil-cpp:absl" ]
   }
-
-  # TODO(b/380940036): remove the circular dependency.
-  allow_circular_includes_from = [ "//media/base" ]
 
   configs += [ "//media:subcomponent_config" ]
 }
